requirements.txt
opencv_python==4.11.0.86
Pillow==11.2.1
transformers==4.51.3
